Cleaned up cruft, added a config file for centralizing secrets, and moved image processing related things into a folder/class of its own
