+++ application/config.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+
+google_api_key = ""
+google_api_secret = ""
