+++ application/panels/rov_panel.py
@@ -5,8 +5,6 @@ from panels.image_panel import ImageStreamPanel
 from panels.joystick import JoystickPanel
 from panels.networking_panel import NetworkingControlPanel
 
-global container_id
-
 class ROVPanel(wx.Panel):
 
     def __init__(self, parent, container_id=None, log=None):
@@ -15,14 +13,11 @@ class ROVPanel(wx.Panel):
         self.log = log
 
         panel = wx.Panel(self)
-        panel_vbox = wx.BoxSizer(wx.VERTICAL)
 
+        panel_vbox = wx.BoxSizer(wx.VERTICAL)
         websocket_and_networking_hbox = wx.BoxSizer(wx.HORIZONTAL)
 
-        ######################################
-        ## Websocket status information panel
-        ######################################
-
+        # Websocket status information panel
         websocket_status_static_box = wx.StaticBox(panel, -1, 'Websocket Status')
         websocket_status_sizer = wx.StaticBoxSizer(websocket_status_static_box, wx.HORIZONTAL)
 
@@ -48,32 +43,20 @@ class ROVPanel(wx.Panel):
         websocket_status_sizer.Add(ws_joystick_sizer, 0, wx.ALL | wx.TOP, 5)
         websocket_status_sizer.Add(ws_camera_sizer, 0, wx.ALL | wx.TOP, 5)
 
-        ######################################
-        ## Networking panel
-        ######################################
-
+        # Networking panel
         self.networking_io = NetworkingControlPanel(parent=panel, label="Networking", id=-1)
 
-        ######################################
-        ## Image panel
-        ######################################
-
+        # Image panel
         self.image_viewer = ImageStreamPanel(parent=panel, label="Vision Processing", id=-1)
 
-        ######################################
-        ## Joystick panel
-        ######################################
-
+        # Joystick panel
         # TODO: Refactor to reduce to singular call
         joystick_controls = wx.StaticBox(panel, -1, 'Joystick Controls:')
         joystick_main_sizer = wx.StaticBoxSizer(joystick_controls, wx.VERTICAL)
         joystick_panel = JoystickPanel(panel, self.log)
         joystick_main_sizer.Add(joystick_panel)
 
-        ######################################
-        ## Panel construction
-        ######################################
-
+        # Panel construction
         websocket_and_networking_hbox.Add(websocket_status_sizer, 0, wx.ALL | wx.CENTER, 5)
         websocket_and_networking_hbox.Add(self.networking_io.panel, 0, wx.ALL | wx.CENTER, 5)
         panel_vbox.Add(websocket_and_networking_hbox, 0, wx.ALL | wx.CENTER, 5)
