+++ requirements.txt
@@ -18,3 +18,4 @@ txaio==18.8.1
 wxPython==4.0.6
 zope.interface==4.6.0
 pydocker
+google-cloud-automl
